feat: refactor account menu to use FormattedAddress component and add useNNSName hook for address resolution
